requirements.txt
streamlit>=1.18.0
crewai>=0.11.0
python-dotenv>=1.0.0
